Don't use 'IMarshal.Vftbl' on .NET
